fix(orchestrator): expose and decompose operator controls
